ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.2
mne==1.8.0
PyWavelets==1.7.0
onnxruntime==1.19.2
torch==2.4.1
--extra-index-url https://download.pytorch.org/whl/cpu
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
skl2onnx==1.17.0
brainflow==5.12.1
websockets==12.0
joblib>=1.3.0
openneuro-py>=2024.2.0
moabb>=1.1.0
kaggle>=1.6.0
kagglehub>=0.3.0
lightgbm>=4.0.0
h5py>=3.0.0
pandas>=2.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
ddtrace>=2.0.0
